main: Add parseArguments function
When main.py is ran from the command line via shell, an argument such as '*.avi'
is already 'globbed', so all the files that match that pattern are
passed as arguments. This is not the case when using an IDE to run it.
parseArguments makes main.py run correctly in both cases.